requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
arxiv
opencv-python-headless
numpy==1.26.4
Pillow
python-multipart